requirements.txt
pydantic
pandas
gunicorn
joblib
scikit-learn
sklearn.preprocessing
flask
numpy
